Move controller logic & add separated services logic
